reverse engineered the model loading algorithm; imported the vehicle model environment initialization; improved the texture and texDictionary handling (now not using GTA:SA slots anymore); fixed camera functions (now not crashing for EU build anymore); added per-entity collisions (WIP); cleaned up code-base (removed unused prototypes);

+++ MTA10/mods/shared_logic/lua/CLuaArguments.cpp
@@ -311,7 +311,7 @@ static inline bool RakNet_WriteArgumentByType( NetBitStreamInterface& stream, lu
         type.data.ucType = LUA_TNUMBER;
         stream.Write( &type );
 
-        num = lua_tonumber( L, idx );
+        num = (float)lua_tonumber( L, idx );
         iNum = (int)num;
 
         if ( num == iNum )
